src/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

azure-identity==1.17.1
azure-ai-projects==1.0.0
azure-ai-agents==1.0.0
